Switch time limits to milliseconds (i32)
Change time limit representation from floating-point seconds (f32) to integer milliseconds (i32) across the codebase. Updated Task, CppRunner, and exec_runner interfaces and fields (including docs and default in Task::new), adjusted timer invocation to pass integer milliseconds, and updated tests to use millisecond values. Also trimmed/cleaned up ANALYSIS.md accordingly. This avoids float rounding for time limits and simplifies passing values to the timer utility.

+++ src/runner/cpp_runner.rs
@@ -23,7 +23,7 @@ pub struct ProgramHandle {
 struct Task {
     program: ProgramHandle,
     input: String,
-    time_limit: f32, // in seconds
+    time_limit: i32, // in milliseconds
     result: Option<RunResult>,
 }
 
@@ -131,7 +131,7 @@ impl CppRunner {
     /// * `program` - Handle to the executable to run.
     /// * `input` - Data to be sent to stdin.
     /// * `time_limit` - Maximum CPU time in seconds.
-    fn add_task(&mut self, program: ProgramHandle, input: String, time_limit: f32) -> TaskHandle {
+    fn add_task(&mut self, program: ProgramHandle, input: String, time_limit: i32) -> TaskHandle {
         trace!("Adding task for program id: {}, time limit: {}", program.id, time_limit);
         let handle = TaskHandle { id: self.tasks.len() };
         self.tasks.push(Task {
@@ -160,7 +160,7 @@ impl CppRunner {
     /// Runs multiple programs against a single input sequentially or in parallel.
     ///
     /// This is a convenience method that manages task creation and result collection.
-    pub fn check_programs(&mut self, input: &str, programs: &[ProgramHandle], time_limit: f32) -> Result<Vec<RunResult>> {
+    pub fn check_programs(&mut self, input: &str, programs: &[ProgramHandle], time_limit: i32) -> Result<Vec<RunResult>> {
         self.clear_tasks();
         let mut handles = Vec::new();
         for &program in programs {
